refactor: renamed tasklist to draggableTaskList
