+++ src/page/backlog/useBacklog.ts
@@ -9,7 +9,7 @@ import {
   saveDatedTasks,
   storeBacklogTasks,
 } from "../../services/tasks";
-import { DraggableTaskProps } from "../../components/taskList";
+import { DraggableTaskProps } from "../../components/draggableTaskList";
 import { DropResult } from "react-beautiful-dnd";
 
 export function useBacklog() {
@@ -87,8 +87,15 @@ export function useBacklog() {
 
   function onDragEnd(result: DropResult) {
     const { destination, source } = result;
+    console.log("🚀 ~ onDragEnd ~ destination, source:", destination, source);
 
     if (!destination) return;
+    if (
+      source.droppableId !== "backlogList" ||
+      destination.droppableId !== "backlogList"
+    )
+      return;
+
     const newTaskArray = [...backlogTasks];
     const [draggedItem] = newTaskArray.splice(source.index, 1);
     newTaskArray.splice(destination.index, 0, draggedItem);
